fixed topology sampling
Former-commit-id: 439291cc8b648c790e32286a4bdbef75fd98b3b0

+++ examples/exahype2/swe/TopologyParser.h
@@ -24,11 +24,11 @@ class examples::exahype2::swe::TopologyParser {
     std::string filepath;
     int ncols = -1; 
     int nrows = -1;
-    float xlowerleft = 0.0;
-    float ylowerleft = 0.0;
-    float cellsize = 0.0;
-    float nodatavalue = 0.0;   
-    std::vector <float> zvalues;
+    double xlowerleft = 0.0;
+    double ylowerleft = 0.0;
+    double cellsize = 0.0;
+    double nodatavalue = 0.0;   
+    std::vector <double> zvalues;
 
   public:
   
@@ -36,16 +36,16 @@ class examples::exahype2::swe::TopologyParser {
     ~TopologyParser();
   
     void parsetopofile();
-    float sampletopology(float x, float y);
+    double sampletopology(double x, double y);
     
     std::string getfilepath() {return this->filepath;}
     int getncols() {return this->ncols;} 
     int getnrows() {return this->nrows;}
-    float getxlowerleft() {return this->xlowerleft;}
-    float getylowerleft() {return this->ylowerleft;}
-    float getcellsize() {return this->cellsize;}
-    float getnodatavalue() {return this->nodatavalue;}   
-    std::vector <float> getzvalues() {return this->zvalues;}
+    double getxlowerleft() {return this->xlowerleft;}
+    double getylowerleft() {return this->ylowerleft;}
+    double getcellsize() {return this->cellsize;}
+    double getnodatavalue() {return this->nodatavalue;}   
+    std::vector <double> getzvalues() {return this->zvalues;}
     
             
 };
